task_2_additional/requirements.txt
selenium==2.53.6
page_objects
pyaml